mobile footer and get rid of extra white space
