+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import "@mantine/core/styles.layer.css";
 import { MantineProvider, AppShell } from "@mantine/core";
 import Navbar from "@/components/Navbar/Navbar";
 import SideNavMenu from "@/components/Navbar/SideNavMenu";
-import Footer from "@/components/Footer/Footer"
+import Footer from "@/components/Footer/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -46,11 +46,11 @@ export default function RootLayout({ children }) {
               <SideNavMenu links={links} />
             </AppShell.Navbar>
 
-            <AppShell.Main>{children}</AppShell.Main>
+            <AppShell.Main style={{ paddingBottom: "0" }}>
+              {children}
+            </AppShell.Main>
 
-            <AppShell.Footer style={{ position: "absolute" }}>
-              {<Footer></Footer>}
-            </AppShell.Footer>
+            <Footer></Footer>
           </AppShell>
         </MantineProvider>
       </body>
